+++ activerecord/lib/active_record/tasks/database_tasks.rb
@@ -3,10 +3,41 @@ module ActiveRecord
     class DatabaseAlreadyExists < StandardError; end # :nodoc:
     class DatabaseNotSupported < StandardError; end # :nodoc:
 
-    module DatabaseTasks # :nodoc:
+    # <tt>ActiveRecord::Tasks::DatabaseTasks</tt> is a utility class, which encapsulates
+    # logic behind common tasks used to manage database and migrations.
+    #
+    # The tasks defined here are used in rake tasks provided by Active Record.
+    #
+    # In order to use DatabaseTasks, a few config values need to be set. All the needed
+    # config values are set by rails already, so it's necessary to do it only if you
+    # want to change the default or when you want to use Active Record outside of Rails
+    # (in such case after cofniguring database tasks, you can also use rake tasks defined in
+    # Active Record)
+    #
+    #
+    # The possible config values are the following:
+    #
+    #   * env: current environment
+    #   * db_dir: your 'db' directory
+    #   * seed_loader: an object which will load seeds, it needs to respond to `load_seed` method
+    #   * database_configuration: configuration of your databases (as in config/database.yml)
+    #   * migrations_paths: a list of paths to directories with migrations
+    #   * fixtures_path: a path to fixtures directory
+    #
+    # Example usage of DatabaseTasks outside Rails could look as such:
+    #
+    #     include ActiveRecord::Tasks
+    #     DatabaseTasks.database_configuration = YAML.load(File.read('my_database_config.yml'))
+    #     DatabaseTasks.db_dir = 'db'
+    #     # other settings...
+    #
+    #     DatabaseTasks.create_current('production')
+    module DatabaseTasks
       extend self
 
       attr_writer :current_config
+      attr_accessor :database_configuration, :migrations_paths, :seed_loader, :db_dir,
+                    :fixtures_path, :env
 
       LOCAL_HOSTS    = ['127.0.0.1', 'localhost']
 
@@ -24,7 +55,7 @@ module ActiveRecord
       register_task(/(oci|oracle)/, ActiveRecord::Tasks::OracleDatabaseTasks)
 
       def current_config(options = {})
-        options.reverse_merge! :env => Rails.env
+        options.reverse_merge! :env => env
         if options.has_key?(:config)
           @current_config = options[:config]
         else
@@ -50,7 +81,7 @@ module ActiveRecord
         each_local_configuration { |configuration| create configuration }
       end
 
-      def create_current(environment = Rails.env)
+      def create_current(environment = env)
         each_current_configuration(environment) { |configuration|
           create configuration
         }
@@ -73,7 +104,7 @@ module ActiveRecord
         each_local_configuration { |configuration| drop configuration }
       end
 
-      def drop_current(environment = Rails.env)
+      def drop_current(environment = env)
         each_current_configuration(environment) { |configuration|
           drop configuration
         }
@@ -83,7 +114,7 @@ module ActiveRecord
         drop database_url_config
       end
 
-      def charset_current(environment = Rails.env)
+      def charset_current(environment = env)
         charset ActiveRecord::Base.configurations[environment]
       end
 
@@ -92,7 +123,7 @@ module ActiveRecord
         class_for_adapter(configuration['adapter']).new(*arguments).charset
       end
 
-      def collation_current(environment = Rails.env)
+      def collation_current(environment = env)
         collation ActiveRecord::Base.configurations[environment]
       end
 
@@ -117,6 +148,16 @@ module ActiveRecord
         class_for_adapter(configuration['adapter']).new(*arguments).structure_load(filename)
       end
 
+      def load_seed
+        if seed_loader
+          seed_loader.load_seed
+        else
+          raise "You tried to load seed data, but no seed loader is specified. Please specify seed " +
+                "loader with ActiveRecord::Tasks::DatabaseTasks.seed_loader = your_seed_loader\n" +
+                "Seed loader should respond to load_seed method"
+        end
+      end
+
       private
 
       def database_url_config
@@ -134,7 +175,7 @@ module ActiveRecord
 
       def each_current_configuration(environment)
         environments = [environment]
-        environments << 'test' if environment.development?
+        environments << 'test' if environment == 'development'
 
         configurations = ActiveRecord::Base.configurations.values_at(*environments)
         configurations.compact.each do |configuration|